backend/requirements.txt
ultralytics
deskew
paddleocr
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python<=4.6.0.66
opencv-contrib-python<=4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
PyMuPDF<1.21.0
Pillow>=10.0.0
pyyaml
einops
paddlepaddle-gpu
gradio